GBIW-253 Restant van filtercode afgsplitst van attributelist-table.component
