+++ tailormap-components/projects/core/src/lib/user-interface/attributelist/attributelist-common/attributelist-filter.ts
@@ -1,6 +1,34 @@
-import { LayerFilterValues } from './attributelist-filter-models';
+import {
+  FilterColumns,
+  FilterValueSettings,
+  LayerFilterValues,
+} from './attributelist-filter-models';
+import {
+  UniqueValuesResponse,
+  ValueParameters,
+} from '../../../shared/value-service/value-models';
+import {
+  MatDialog,
+  MatDialogConfig,
+} from '@angular/material/dialog';
+import { AttributeDataSource } from './attributelist-datasource';
+import { AttributelistFilterValuesFormComponent } from '../attributelist-filter-values-form/attributelist-filter-values-form.component';
+import { ValueService } from '../../../shared/value-service/value.service';
+import { AttributelistRefresh } from './attributelist-models';
 
 export class AttributelistFilter {
+
+  constructor(
+    private dataSource: AttributeDataSource,
+    private valueService: ValueService,
+    private dialog: MatDialog,
+  ) {
+  }
+
+  private valueParams: ValueParameters = {
+    applicationLayer: 0,
+    attributes: [],
+  }
 
   public layerFilterValues: LayerFilterValues = {
     layerId: 0,
@@ -8,6 +36,17 @@ export class AttributelistFilter {
   };
 
   private valueFilter: string;
+
+  public initFiltering(colNames: string[]): void {
+    // Init the filter structure
+    this.layerFilterValues.layerId = this.dataSource.params.layerId;
+    // const colNames = this.getColumnNames();
+    for (const colName of colNames) {
+      let filterColumn: FilterColumns;
+      filterColumn = {name: colName, status: false, nullValue: false, uniqueValues: []};
+      this.layerFilterValues.columns.push(filterColumn);
+    }
+  }
 
   /**
    * Create the CQL filter string
@@ -49,4 +88,55 @@ export class AttributelistFilter {
     return this.valueFilter;
   }
 
+  public setFilter(attributelistRefresh: AttributelistRefresh, columnName: string): void {
+    // Get the unique values for this column
+    this.valueParams.applicationLayer = this.dataSource.params.layerId;
+    this.valueParams.attributes = [];
+    this.valueParams.attributes.push(columnName);
+    this.valueService.uniqueValues(this.valueParams).subscribe((data: UniqueValuesResponse) => {
+      if (data.success) {
+        let uniqueValues: FilterValueSettings[];
+        uniqueValues = [];
+        const colObject = this.layerFilterValues.columns.find(c => c.name === columnName);
+        const colIndex = this.layerFilterValues.columns.findIndex(obj => obj.name === columnName);
+        if (colObject.uniqueValues.length === 0) {
+
+          data.uniqueValues[columnName].forEach(val => {
+            let filterValueSettings: FilterValueSettings;
+            filterValueSettings = {value: val, select: true};
+            uniqueValues.push(filterValueSettings);
+          })
+        } else {
+          colObject.uniqueValues.forEach(val => uniqueValues.push(Object.assign({}, val)))
+        }
+
+        const config = new MatDialogConfig();
+        config.data = {
+          colName: columnName,
+          values: uniqueValues,
+        };
+        const dialogRef = this.dialog.open(AttributelistFilterValuesFormComponent, config);
+        dialogRef.afterClosed().subscribe(filterSetting => {
+          // Do the filtering
+          if (filterSetting !== 'CANCEL') {
+            if (filterSetting === 'ON') {
+              this.layerFilterValues.columns[colIndex].uniqueValues = config.data.values;
+              this.layerFilterValues.columns[colIndex].nullValue = false;
+              this.layerFilterValues.columns[colIndex].status = true;
+            } else if (filterSetting === 'NONE') {
+              this.layerFilterValues.columns[colIndex].uniqueValues = config.data.values;
+              this.layerFilterValues.columns[colIndex].nullValue = true;
+              this.layerFilterValues.columns[colIndex].status = true;
+            } else if (filterSetting === 'OFF') {
+              this.layerFilterValues.columns[colIndex].uniqueValues = [];
+              this.layerFilterValues.columns[colIndex].nullValue = false;
+              this.layerFilterValues.columns[colIndex].status = false;
+            }
+            this.dataSource.params.valueFilter = this.createFilter();
+            attributelistRefresh.refreshTable();
+          }
+        });
+      }
+    });
+  }
 }
