+++ tailormap-components/projects/core/src/lib/user-interface/attributelist/attributelist-common/attributelist-models.ts
@@ -3,6 +3,10 @@ export interface AttributelistTable {
   onAfterLoadData: () => void;
 }
 
+export interface AttributelistRefresh {
+  refreshTable: () => void;
+}
+
 export interface AttributelistConfig {
   pageSize: number;
 }
